# backend/docs/swagger.yaml
definitions:
...
  schemas.AuthenticationURL:
    properties:
      authentication_url:
        type: string
    type: object
  schemas.CodeCredentials:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  schemas.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  schemas.JWT:
    properties:
      token:
        type: string
    type: object
...
  schemas.Service:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        $ref: '#/definitions/schemas.ServiceName'
      oauth:
        type: boolean
      update_at:
        type: string
    required:
    - color
    - description
    - icon
    - name
    - oauth
    type: object
  schemas.ServiceName:
    enum:
    - Spotify
    - OpenWeatherMap
    - Timer
    - Gmail
    - Github
    - Dropbox
    - Microsoft
    type: string
    x-enum-varnames:
    - Spotify
    - Openweathermap
    - Timer
    - Gmail
    - Github
    - Dropbox
    - Microsoft
  schemas.Token:
    properties:
      created_at:
        description: Time when the token was created
        type: string
      expire_at:
        description: Time when the token expires
        type: string
      id:
        description: Unique identifier for the token
        type: integer
      refresh_token:
        description: Refresh token
        type: string
      service:
        allOf:
        - $ref: '#/definitions/schemas.Service'
        description: Service that the token belongs to
      token:
        description: Token
        type: string
      update_at:
        description: Time when the token was last updated
        type: string
      user:
        allOf:
        - $ref: '#/definitions/schemas.User'
        description: User that the token belongs to
    type: object
  schemas.User:
    properties:
      created_at:
        description: Time when the user was created
        type: string
      email:
        description: Email of the user
        type: string
      id:
        description: Unique identifier for the user
        type: integer
      password:
        description: can be null for Oauth2.0 users
        type: string
      token_id:
        description: Foreign key for LinkURL
        type: integer
      updated_at:
        description: Time when the user was last updated
        type: string
      username:
        description: Username of the user
        type: string
    required:
    - email
    - username
    type: object
  schemas.UserAllInfo:
    properties:
      tokens:
        description: List of tokens
        items:
          $ref: '#/definitions/schemas.Token'
        type: array
      user:
        allOf:
        - $ref: '#/definitions/schemas.User'
        description: User
    type: object
  schemas.UserCredentials:
    properties:
      email:
        description: Email of the user
        type: string
      username:
        description: Username of the user
        type: string
    type: object
info:
  contact: {}
paths:
...
  /microsoft/auth:
    get:
      consumes:
      - application/json
      description: give url to authenticate with microsoft
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AuthenticationURL'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with microsoft
      tags:
      - Microsoft
  /microsoft/auth/callback:
    post:
      consumes:
      - application/json
      description: give url to authenticate with microsoft
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: give url to authenticate with microsoft
      tags:
      - Microsoft
  /microsoft/auth/callback/mobile:
    post:
      consumes:
      - application/json
      description: give url to authenticate with microsoft
      parameters:
      - description: Callback Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CodeCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give url to authenticate with microsoft
      tags:
      - Microsoft
  /microsoft/info:
    get:
      consumes:
      - application/json
      description: give user info of microsoft
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserCredentials'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of microsoft
      tags:
      - Microsoft
...
  /service/info:
    get:
      consumes:
      - application/json
      description: get service info of service id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.Service'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: get service info
      tags:
      - Service
  /service/info/:id:
    get:
      consumes:
      - application/json
      description: get service info of service id
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Service'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: get service info
      tags:
      - Service
...
  /user/info:
    get:
      consumes:
      - application/json
      description: give user info of user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserCredentials'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of user
      tags:
      - User
  /user/info/:
    delete:
      consumes:
      - application/json
      description: give user info of user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: give user info of user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of user
      tags:
      - User
  /user/info/all:
    get:
      consumes:
      - application/json
      description: give user info of user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserAllInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      security:
      - bearerAuth: []
      summary: give user info of user
      tags:
      - User
  /user/login:
    post:
      description: Authenticates a user and provides a JWT to Authorize API calls
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.JWT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Provides a JSON Web Token
      tags:
      - User
  /user/register:
    post:
      description: Authenticates a user and provides a JWT to Authorize API calls
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.JWT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.ErrorResponse'
      summary: Provides a JSON Web Token
      tags:
      - User
securityDefinitions:
  bearerAuth:
    description: Use "Bearer <token>" as the format for the Authorization header.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"